backend/requirements.txt
inferedge-moss>=1.0.0b15
python-dotenv>=1.2.1
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
